ECDsa key based on NCrypt
This fixes key importing and gets the last test (PerfectForwardSecrecy) working.
